+++ packages/tendermint-rpc/src/index.ts
@@ -20,9 +20,7 @@ export {
   ValidatorSecp256k1Pubkey,
   ValidatorPubkey,
 } from "./types";
-export * as tendermint33 from "./tendermint33";
 export {
-  Tendermint33Client,
   AbciInfoResponse,
   AbciQueryResponse,
   Attribute,
@@ -89,6 +87,6 @@ export {
   TxSearchRequest,
   ValidatorsRequest,
   ValidatorsParams,
-} from "./tendermint33";
+} from "./tendermint34";
 export * as tendermint34 from "./tendermint34";
 export { Tendermint34Client } from "./tendermint34";